+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## UNRELEASED
+## 1.0.3 (January 30, 2023)
 
 IMPROVEMENTS:
 * Helm: